+++ userland/examples/nrf51_aes_ccm/aes.c
